royal-brain/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
email-validator
python-jose[cryptography]
argon2-cffi
python-multipart
openai
web3
psycopg[binary]

# testing
pytest
httpx
